types: deleted ccf.ts,  moved type to common types, fixed  types' path  in teads-aws
