--- src/types/ccf.ts
@@ -1,27 +1,0 @@
-/**
- * Consumption information for a single instance.
- */
-export type IConsumption = {
-  idle: number;
-  tenPercent: number;
-  fiftyPercent: number;
-  hundredPercent: number;
-  minWatts: number;
-  maxWatts: number;
-};
-
-/**
- * Information about a single compute instance.
- */
-export type IComputeInstance = {
-  consumption: IConsumption;
-  embodiedEmission: number;
-  name: string;
-  vCPUs: number;
-  maxvCPUs: number;
-};
-
-export type ICcfResult = {
-  energy: number;
-  embodied_emissions: number;
-};
