+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './boavizta';
-export * from './ccf';
 export * from './common';
